Refactored

+++ src/constants.hpp
@@ -2,7 +2,7 @@
 #define CONSTANTS_H
 
 namespace constants {
-	constexpr unsigned int kDataLength = 20;
+	constexpr size_t kDataLength = 20;
 	constexpr int kNumCalcFrames = 5;
 	constexpr int kNumPredictedFrames = 5;
 	constexpr int kStartFrame = 20;
